+++ src/entities/order/ui/Card.scss
